implement MergedSortingExtractor
